git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@7719 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/rt/src/chregistry.c
@@ -81,9 +81,9 @@ ROMCONST chdebug_t ch_debug = {
   {'m', 'a', 'i', 'n'},
   (uint8_t)0,
   (uint8_t)sizeof (chdebug_t),
-  (uint16_t)((CH_KERNEL_MAJOR << 11) |
-             (CH_KERNEL_MINOR << 6) |
-             (CH_KERNEL_PATCH << 0)),
+  (uint16_t)(((unsigned)CH_KERNEL_MAJOR << 11U) |
+             ((unsigned)CH_KERNEL_MINOR << 6U) |
+             ((unsigned)CH_KERNEL_PATCH << 0U)),
   (uint8_t)sizeof (void *),
   (uint8_t)sizeof (systime_t),
   (uint8_t)sizeof (thread_t),
